added loggedUserMiddleware
